reworking test-support module
