+++ modules/network/pom.xml
@@ -22,6 +22,12 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
+			<groupId>hosh</groupId>
+			<artifactId>hosh-spi-test-support</artifactId>
+			<version>${project.version}</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
 			<groupId>jakarta.json</groupId>
 			<artifactId>jakarta.json-api</artifactId>
 			<version>2.0.0-RC2</version>
